Refactored SparklineContainer & Dialog to use one query
