+++ src/components/ProjectList/CodeClimate/LetterGradeContainer.jsx
@@ -1,37 +1,28 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Maintainability } from './letterGrade.module.scss';
 import LetterGrade from './LetterGrade';
 import getColor from '../../../utils/getColorFromCCGrade';
-import ChartDialog from '../GitHub/Charts/ChartDialog'
+
 
 const LetterGradeContainer = props => {
-  const [currentIndex, setCurrentIndex] = useState(0)
-  const [dialogOpen, setDialogOpen] = useState({
-    'sparkChart': false,
-  })
+  // const [currentIndex, setCurrentIndex] = useState(0)
+  // const [dialogOpen, setDialogOpen] = useState({
+  //   'sparkChart': false,
+  // })
 
   if (!Array.isArray(props.ghrepos) || !props.ghrepos.length) return null;
   
-  const toggleDialog = name => {
-    setDialogOpen({ ...dialogOpen, [name]: !dialogOpen[name] })
-  }
+  // const toggleDialog = name => {
+  //   setDialogOpen({ ...dialogOpen, [name]: !dialogOpen[name] })
+  // }
 
-  const handleSparkClick = (repoIndex) => {
-    toggleDialog('sparkChart')
-    setCurrentIndex(repoIndex)
-  }
-  
+  // const handleSparkClick = (repoIndex) => {
+  //   toggleDialog('sparkChart')
+  //   setCurrentIndex(repoIndex)
+  // }
+
   return (
     <>
-      <ChartDialog
-        ghrepos={props.ghrepos}
-        name={props.ghrepos[currentIndex].name}
-        open={dialogOpen['sparkChart']}
-        toggleDialog={toggleDialog}
-        onClose={() => toggleDialog('sparkChart')}
-        projectName={props.name}
-        setCurrentIndex={setCurrentIndex}
-        currentIndex={currentIndex} />
       <div className={Maintainability}>
         {props.ghrepos.map((repo, index) => {
           const color = getColor(repo.grade);
@@ -41,10 +32,11 @@ const LetterGradeContainer = props => {
             color={color}
             repo={repo}
             repoIndex={index}
-            currentIndex={currentIndex}
-            setCurrentIndex={setCurrentIndex}
-            projectName={props.name}
-            handleSparkClick={handleSparkClick} />
+            // currentIndex={currentIndex}
+            // setCurrentIndex={setCurrentIndex}
+            // projectName={props.name}
+            // handleSparkClick={handleSparkClick}
+             />
         })}
       </div>
     </>
